# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Max Hackathon API",
    "description": "API «Навигатора мер поддержки»: каталог программ, профиль пользователя, подбор, «Мои заявки» и ИИ-консультант. Его используют бот и mini app в MAX.\n\nЭндпоинты с данными пользователя требуют аутентификации одним из способов:\n- `Authorization: tma <WebApp.initData>` — mini app (подпись HMAC-SHA256 на токене бота);\n- `X-Service-Token` + `X-Max-User-Id` — бот и проверка жюри.\n\nПользователь определяется только по аутентификации — прочитать чужие данные подстановкой id нельзя.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/applications/": {
      "post": {
        "tags": [
          "Applications"
        ],
        "summary": "Create Application",
        "description": "Save a program into \"Мои заявки\" for the caller (idempotent).",
        "operationId": "create_application_api_v1_applications__post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`tma <WebApp.initData>` — подписанные данные запуска mini app",
              "title": "Authorization"
            },
            "description": "`tma <WebApp.initData>` — подписанные данные запуска mini app"
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Сервисный токен (SERVICE_TOKEN) для вызовов бота и проверки жюри",
              "title": "X-Service-Token"
            },
            "description": "Сервисный токен (SERVICE_TOKEN) для вызовов бота и проверки жюри"
          },
          {
            "name": "x-max-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID пользователя MAX; вместе с X-Service-Token",
              "title": "X-Max-User-Id"
            },
            "description": "ID пользователя MAX; вместе с X-Service-Token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            }
          },
          "401": {
            "description": "Нет или неверная аутентификация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Сервисный токен без X-Max-User-Id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Applications"
        ],
        "summary": "List Applications",
        "description": "The caller's saved programs, newest first.",
        "operationId": "list_applications_api_v1_applications__get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`tma <WebApp.initData>` — подписанные данные запуска mini app",
              "title": "Authorization"
            },
            "description": "`tma <WebApp.initData>` — подписанные данные запуска mini app"
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Сервисный токен (SERVICE_TOKEN) для вызовов бота и проверки жюри",
              "title": "X-Service-Token"
            },
            "description": "Сервисный токен (SERVICE_TOKEN) для вызовов бота и проверки жюри"
          },
          {
            "name": "x-max-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID пользователя MAX; вместе с X-Service-Token",
              "title": "X-Max-User-Id"
            },
            "description": "ID пользователя MAX; вместе с X-Service-Token"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationRead"
                  },
                  "title": "Response List Applications Api V1 Applications  Get"
                }
              }
            }
          },
          "401": {
            "description": "Нет или неверная аутентификация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Сервисный токен без X-Max-User-Id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/{application_id}": {
      "patch": {
        "tags": [
          "Applications"
        ],
        "summary": "Update Application",
        "description": "Move a saved program along the saved → in_progress → submitted track,\nand/or record which checklist documents are ready.",
        "operationId": "update_application_api_v1_applications__application_id__patch",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`tma <WebApp.initData>` — подписанные данные запуска mini app",
              "title": "Authorization"
            },
            "description": "`tma <WebApp.initData>` — подписанные данные запуска mini app"
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Сервисный токен (SERVICE_TOKEN) для вызовов бота и проверки жюри",
              "title": "X-Service-Token"
            },
            "description": "Сервисный токен (SERVICE_TOKEN) для вызовов бота и проверки жюри"
          },
          {
            "name": "x-max-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID пользователя MAX; вместе с X-Service-Token",
              "title": "X-Max-User-Id"
            },
            "description": "ID пользователя MAX; вместе с X-Service-Token"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationRead"
                }
              }
            }
          },
          "401": {
            "description": "Нет или неверная аутентификация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Сервисный токен без X-Max-User-Id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Объект не найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/applications/by-program/{program_id}": {
      "delete": {
        "tags": [
          "Applications"
        ],
        "summary": "Delete Application By Program",
        "description": "Un-save a program. Keyed by program id because that is what the\nmini-app's bookmark toggle knows about.",
        "operationId": "delete_application_by_program_api_v1_applications_by_program__program_id__delete",
        "parameters": [
          {
            "name": "program_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Program Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`tma <WebApp.initData>` — подписанные данные запуска mini app",
              "title": "Authorization"
            },
            "description": "`tma <WebApp.initData>` — подписанные данные запуска mini app"
          },
          {
            "name": "x-service-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Сервисный токен (SERVICE_TOKEN) для вызовов бота и проверки жюри",
              "title": "X-Service-Token"
            },
            "description": "Сервисный токен (SERVICE_TOKEN) для вызовов бота и проверки жюри"
          },
          {
            "name": "x-max-user-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID пользователя MAX; вместе с X-Service-Token",
              "title": "X-Max-User-Id"
            },
            "description": "ID пользователя MAX; вместе с X-Service-Token"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Нет или неверная аутентификация",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Сервисный токен без X-Max-User-Id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApplicationCreate": {
        "properties": {
          "program_id": {
            "type": "integer",
            "title": "Program Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "saved"
          }
        },
        "type": "object",
        "required": [
          "program_id"
        ],
        "title": "ApplicationCreate",
        "description": "Save a program into \"Мои заявки\".\n\nThe profile is resolved from the authenticated caller, so the body only\nnames the program."
      },
      "ApplicationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "profile_id": {
            "type": "integer",
            "title": "Profile Id"
          },
          "program_id": {
            "type": "integer",
            "title": "Program Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "checked_docs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Checked Docs",
            "default": []
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "program": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SupportProgramRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "profile_id",
          "program_id",
          "status",
          "created_at"
        ],
        "title": "ApplicationRead"
      },
      "ApplicationUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "checked_docs": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked Docs"
          }
        },
        "type": "object",
        "title": "ApplicationUpdate",
        "description": "Partial update: omitted fields are left as they are."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Body of every 4xx the API raises itself (FastAPI's HTTPException)."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SupportProgramRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "short_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Short Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "industries": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industries"
          },
          "eligible_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eligible Status"
          },
          "conditions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "amount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "highlight": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highlight"
          },
          "deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "doc_checklist": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Checklist"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "checked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked At"
          },
          "is_mock": {
            "type": "boolean",
            "title": "Is Mock"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "is_mock",
          "created_at",
          "updated_at"
        ],
        "title": "SupportProgramRead"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "servers": [
    {
      "url": "https://maxnalog.ru:8444",
      "description": "Продакшен (проверка жюри)"
    },
    {
      "url": "http://localhost:8000",
      "description": "Локально, docker compose up"
    }
  ]
}